If the human-readable print is uncommented, the code could crash
The potential bug is "use after free".

If `message` is freed before being used, depending on the local
implementation of `free`, the pointers within it might be zeroed or be
otherwise made invalid, and `asn_fprint` would then crash trying to
dereference those pointers. This change moves freeing of the buffer to
after its final use.