Launch robot_state_publisher to publish transforms for CARMA Messenger vehicle (#186)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

<!--- Describe your changes in detail -->
On CARMA Messenger vehicles, velocity is received by subscribing to the
/position/velocity topic, which is published by the
pin_point_application node. However, in order for the
pin_point_application node to publish velocity, a transform is first
conducted between the pinpoint sensor frame and the base_link frame. For
this transform to be available, the transform information must be
available on the /tf_static topic.

In this PR, updates have been made to carma-messenger's launch files to
launch the robot_state_publisher node and pass it the urdf file (from
the Chevrolet Tahoe's vehicle calibration directory), which enables the
node to publish the transform information. These updates are modeled
after a similar implementation in carma-platform for publishing urdf
information.

Related carma-torc-pinpoint-driver PR: [PR
#33](https://github.com/usdot-fhwa-stol/carma-torc-pinpoint-driver/pull/33)
Related carma-vehicle-calibration PR: [PR
#113](https://github.com/usdot-fhwa-stol/carma-vehicle-calibration/pull/113)

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->
Issue #187 

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
This change is required since the new emergency_response_vehicle_plugin
in CARMA Messenger must subscribe to the /position/velocity topic to
receive the current vehicle velocity.

## How Has This Been Tested?
Local integration test.

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.