Car 5924 Migrate ROS Bridge to ROS2 (#227)
# PR Details
## Description

This PR migrates the ROS Bridge used in the CARMA Messenger Web UI to
the ROS2 version. Additional modifications were made to the Web UI to
ensure that it now receives ROS2 messages.

## Related Jira Key

[CAR-5924](https://usdot-carma.atlassian.net/browse/CAR-5924)

## Motivation and Context

The rosbridge package (which enables communication between the UI and
the ROS System) was upgraded from ROS 1 Noetic to ROS 2 Foxy for
carma-platform, but was never upgraded for carma-messenger. The goal of
this story is to complete that upgrade for carma-messenger.

## How Has This Been Tested?

A local integration test was conducted to verify that the Web UI still
functions as expected when launching CARMA Messenger. The docker
containers were inspected to ensure that only the ROS2 version of the
ROS bridge is launched and that messages published in ROS2 by the UI
were received in the ROS1 container.

## Types of changes

- [ ] Defect fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.


[CAR-5924]:
https://usdot-carma.atlassian.net/browse/CAR-5924?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ