Fix/increase start end loc timeout (#282)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

This PR increases the timeout threshold for resetting the start and end
position of the work zone created using the CARMA Messenger UI. The
start and end zone locations are configured to publish at 10 Hz.
However, if the browser is changed to a different page, the publishing
rate drops from 10 Hz to ~1 Hz. Increasing the timeout ensures that we
continue to use the latest received start and end zone positions when
this happens.

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key

[CDAD-186](https://usdot-carma.atlassian.net/browse/CDAD-186)

## Motivation and Context

Needed for the traffic_incident node to continue publishing the correct
workzone when the tab is switched.
## How Has This Been Tested?

Tested locally and confirmed the MOM continue to use the correct
coordinates when switching to a different tab.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.


[CDAD-186]:
https://usdot-carma.atlassian.net/browse/CDAD-186?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ